chore: split config and locale
