+++ packages/frontend-embed/src/i18n.ts
@@ -6,7 +6,7 @@
 import { markRaw } from 'vue';
 import { I18n } from '@@/js/i18n.js';
 import type { Locale } from '../../../locales/index.js';
-import { locale } from '@@/js/config.js';
+import { locale } from '@@/js/locale.js';
 
 export const i18n = markRaw(new I18n<Locale>(locale, _DEV_));
