+++ packages/frontend/src/i18n.ts
@@ -5,8 +5,8 @@
 
 import { markRaw } from 'vue';
 import { I18n } from '@@/js/i18n.js';
+import { locale } from '@@/js/locale.js';
 import type { Locale } from '../../../locales/index.js';
-import { locale } from '@@/js/config.js';
 
 export const i18n = markRaw(new I18n<Locale>(locale, _DEV_));
